requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
PyYAML
PyNaCl